feat: split submission components
